Add @feed_items used for an unsuccessful micropost creation
When post a new micropost, the MicropostsController#create will handle
that request. 

If the micropost is created successfully, the controller
will redirect user to homepage, where @feed_items is available. 

If the micropost isn't create successfully, the controller will render
the Homepage to user, but it is still in the MicropostsController, where
there is not any @feed_items available.